refactor: move KV-transfer enums into a backend-agnostic header
